requirements.txt
# Core packages
langchain==0.0.354
langchain-community==0.0.22
langchain-openai==0.0.5

# Embeddings and vector stores
sentence-transformers==2.3.1
faiss-cpu==1.7.4

# Document processing
pypdf==3.17.1
docx2txt==0.8

# API and utilities
openai==1.14.2
python-dotenv==1.0.1
tqdm==4.66.2

# Data processing and visualization
numpy==1.24.4
matplotlib==3.7.3
scikit-learn==1.3.2

# Use pre-built wheel for tiktoken instead of building from source
tiktoken==0.5.1; python_version < '3.13'